chore: simplify backend liveness coverage
